pyproject.toml
[project]
name = "trading-ntrader"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "click>=8.2.1",
    "docker>=7.1.0",
    "fastapi>=0.121.2",
    "greenlet>=3.2.4",
    "jinja2>=3.1.6",
    "nautilus-trader[ib]>=1.190.0",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "mypy>=1.18.1",
    "pandas-stubs>=2.3.2.250827",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-docker>=3.2.3",
    "pytest-forked>=1.6.0",
    "pytest-xdist>=3.6.1",
    "pyyaml>=6.0.2",
    "ruff>=0.13.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-setuptools>=80.9.0.20250822",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "trading: marks tests as trading-related (deselect with '-m \"not trading\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "nautilus: marks tests as nautilus framework-specific",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = [
    "tests_archive",
    "scripts",
]

[[tool.mypy.overrides]]
module = [
    "ibapi",
    "ibapi.*",
]
ignore_missing_imports = true

[tool.ruff]
exclude = [
    "tests/old_tests",
    "tests_archive",
    ".github",
    "alembic",
    ".git",
    ".venv",
    "__pycache__",
    "*.ipynb",
]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
